+++ src/main/java/dev/wgrgwg/somniverse/dream/controller/DreamController.java
@@ -13,7 +13,6 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -22,7 +21,6 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
@@ -76,14 +74,6 @@ public class DreamController {
         return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
     }
 
-    @PreAuthorize("hasAnyAuthority('ADMIN', 'MANAGER')")
-    @DeleteMapping("/admin/{dreamId}")
-    public ResponseEntity<ApiResponseDto<Void>> deleteDreamByAdmin(@PathVariable Long dreamId) {
-        dreamService.deleteDreamByAdmin(dreamId);
-
-        return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
-    }
-
     @GetMapping("/my")
     public ResponseEntity<ApiResponseDto<Page<DreamSimpleResponse>>> getMyDreams(
         @AuthenticationPrincipal CustomUserDetails userDetails, Pageable pageable) {
@@ -116,26 +106,6 @@ public class DreamController {
         @PathVariable Long memberId, Pageable pageable) {
         Page<DreamSimpleResponse> response = dreamService.getPublicDreamsByMember(memberId,
             pageable);
-
-        return ResponseEntity.status(HttpStatus.OK).body(ApiResponseDto.success(response));
-    }
-
-    @PreAuthorize("hasAnyAuthority('ADMIN', 'MANAGER')")
-    @GetMapping("/admin")
-    public ResponseEntity<ApiResponseDto<Page<DreamSimpleResponse>>> getDreamsForAdmin(
-        Pageable pageable, @RequestParam(defaultValue = "false") Boolean includeDeleted) {
-
-        Page<DreamSimpleResponse> response = dreamService.getAllDreamsForAdmin(pageable,
-            includeDeleted);
-
-        return ResponseEntity.status(HttpStatus.OK).body(ApiResponseDto.success(response));
-    }
-
-    @PreAuthorize("hasAnyAuthority('ADMIN', 'MANAGER')")
-    @GetMapping("/admin/{dreamId}")
-    public ResponseEntity<ApiResponseDto<DreamResponse>> getDreamAsAdmin(@PathVariable Long dreamId,
-        @RequestParam(defaultValue = "false") Boolean includeDeleted) {
-        DreamResponse response = dreamService.getDreamAsAdmin(dreamId, includeDeleted);
 
         return ResponseEntity.status(HttpStatus.OK).body(ApiResponseDto.success(response));
     }
